+++ cw-orch-daemon/src/builder.rs
@@ -28,8 +28,6 @@ pub struct DaemonAsyncBuilder {
     pub(crate) deployment_id: Option<String>,
     /// Wallet mnemonic
     pub(crate) mnemonic: Option<String>,
-    /// Warn in case of disabled log
-    pub(crate) check_if_log_disabled: Option<bool>,
 }
 
 impl DaemonAsyncBuilder {
@@ -55,12 +53,6 @@ impl DaemonAsyncBuilder {
         self
     }
 
-    /// Disables warning disabled logs output
-    pub fn no_warning(&mut self) -> &mut Self {
-        self.check_if_log_disabled = Some(false);
-        self
-    }
-
     /// Build a daemon
     pub async fn build(&self) -> Result<DaemonAsync, DaemonError> {
         let chain = self
@@ -82,9 +74,7 @@ impl DaemonAsyncBuilder {
             state,
             sender: Rc::new(sender),
         };
-        if self.check_if_log_disabled.unwrap_or(true) {
-            print_if_log_disabled();
-        }
+        print_if_log_disabled()?;
         Ok(daemon)
     }
 }
@@ -95,7 +85,6 @@ impl From<DaemonBuilder> for DaemonAsyncBuilder {
             chain: value.chain,
             deployment_id: value.deployment_id,
             mnemonic: value.mnemonic,
-            check_if_log_disabled: value.check_if_log_disabled,
         }
     }
 }