+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # cw-orchestrator Changelog
 
+## Unreleased
+
+- Add `MutCwEnv` for manipulating testing environments.
+- Add `BankQuerier` as trait bound to `CwEnv`.
+- Add `WasmCodeQuerier` as trait bound to `CwEnv`.
+- Changed the snapshot feature for better snapshot readability.
+
+## 0.18.2
+
+- Added Snapshot-Testing
+
+## 0.18.1
+
+- Dix : Added daemon flag on the networks import
+
 ## 0.18.0
 
 - Added wallet balance assertions to avoid erroring when the wallet doesn't have enough balance when submitting a daemon transaction
@@ -14,11 +29,12 @@
 - Added `Deploy` to prelude
 - Add ability to provide custom state in `Deploy::set_contracts_state`
 - Breaking change: remove the `&self` dependency for the `Deploy::deployed_state_file_path` method
-- Using `dirs` instead of `shellexpand` for getting the default cw-orch state dir. 
+- Using `dirs` instead of `shellexpand` for getting the default cw-orch state dir.
 - Exposed the state_dir location
 - Added better env variable management by @Kayanski
 - Added message to enable logging if not enabled
 - Removed unused dependencies
+- Added snapshot testing in mock env by @Kayanski (feature flagged)
 
 ## v0.16.4 [20th September 2023]
 
@@ -55,13 +71,13 @@
 - Added helper to modify the chain id of cw-multi-test (Mock)
 - Added a trait to be able to commit any transaction to chain (Protobuf any type)
 - Added min gas and average gas utilization for computing the tx fee
-- Added Install Readme 
+- Added Install Readme
 - Change the state file default location for relative paths `./` --> `~/.cw-orchestrator`
 - Added env variables for customizing experience
 
 ## v0.15.0
 
-- Add `add_balance` function on the `Mock` type. 
+- Add `add_balance` function on the `Mock` type.
 
 ## v0.10.0